+++ src/components/Markdown/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-10-10: Initial version.
+- 2025-10-24: Support render code blocks (using internal rehype plugins). Wrap
+  its children in `Typography.Root`.